+++ mapwize-ui/src/main/java/io/mapwize/mapwizeui/BasePresenter.java
@@ -45,4 +45,5 @@ interface BasePresenter {
 
     MapwizeMap getMapwizeMap();
     boolean onBackButtonPressed();
+    void unselectContent();
 }