+++ mapwize-ui/src/main/java/io/mapwize/mapwizeui/MapwizeUIView.java
@@ -711,6 +711,11 @@ public class MapwizeUIView extends FrameLayout implements BaseUIView, SearchBarV
     }
 
     @Override
+    public void setLanguage(String language) {
+        this.searchResultList.setLanguage(language);
+    }
+
+    @Override
     public void showDirectionError() {
         bottomCardView.showDirectionError();
     }
@@ -958,7 +963,7 @@ public class MapwizeUIView extends FrameLayout implements BaseUIView, SearchBarV
             return true;
         }
         if (infoVisible) {
-            hideInfo();
+            presenter.unselectContent();
             return true;
         }
         return false;